+++ examples/resnet50/migraphx.cpp
@@ -171,7 +171,9 @@ Args getArgs(int argc, char** argv) {
   Args args = parseArgs(argc, argv);
 
   if (args.path_to_model.empty()) {
-    fs::path root{std::getenv("AMDINFER_ROOT")};
+    const auto* root_str = std::getenv("AMDINFER_ROOT");
+    assert(root_str != nullptr);
+    fs::path root{root_str};
     args.path_to_model =
       root / "external/artifacts/onnx/resnet50v2/resnet50-v2-7.onnx";
   }